pyproject.toml
# pyproject.toml for Kontextrument

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Kontextrument"
version = "1.0.0"
authors = [
  { name = "Kai SD", email = "kaisd@vp-project.org" }
]
description = "A desktop/CLI assistant to streamline developing software with Large Language Models (LLMs)."
readme = "README.md"
requires-python = ">=3.7"
license = { file = "LICENSE" } # It's best practice to include a LICENSE file.
keywords = ["llm", "ai", "developer-tools", "gui", "wxpython", "automation", "context-management"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Environment :: X11 Applications",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Text Processing",
]

dependencies = [
    "wxPython>=4.2.3",
    "PyPubSub",
    "pathspec",
    "pywinpty; sys_platform == 'win32'",
    "ptyprocess; sys_platform != 'win32'",
]

[project.urls]
"Homepage" = "https://github.com/VP-Project/kontextrument"
"Bug Tracker" = "https://github.com/VP-Project/kontextrument/issues"
"Repository" = "https://github.com/VP-Project/kontextrument"

[project.scripts]
ktr = "ktr_runner:main"

[tool.setuptools]
# This tells setuptools that your packages are in the 'src' directory.
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
# This automatically finds all packages within the 'src' directory.
where = ["src"]
